+++ src/rosbag_manip/__init__.py
@@ -1,5 +1,7 @@
-from .rosbag_manip import rosbag_manipulation
-from .data_types.OdometryData import OdometryData
+from .command_line import CmdLineInterface
+from .conversion_utils import *
 from .data_types.ImageData import ImageData
 from .data_types.ImuData import ImuData
+from .data_types.OdometryData import OdometryData
 from .data_types.Data import CoordinateFrame
+from .rosbag.Ros2BagWrapper import Ros2BagWrapper
