+++ src/rosbag_manip/command_line.py
@@ -1,0 +1,155 @@
+from collections import defaultdict
+from .data_types.ImageData import ImageData
+from .data_types.OdometryData import OdometryData
+from decimal import Decimal
+import matplotlib.pyplot as plt
+import numpy as np
+from pathlib import Path
+from .rosbag.Ros2BagWrapper import Ros2BagWrapper
+from rosbags.rosbag2 import Reader as Reader2
+from rosbags.rosbag2 import Writer as Writer2
+from scipy.spatial.transform import Rotation as R
+import tqdm
+import yaml
+
+class CmdLineInterface():
+
+    def __init__(self, **kwargs):
+        """
+        Initialize the rosbag manipulator with the provided keyword arguments,
+        which is assumed to be a dictionary containing the manipulation configuration.
+
+        Args:
+            **kwargs: Keyword arguments containing the manipulation configuration, including:
+                - 'input_bag': Path to the input ROS2 bag file.
+                - 'output_bag': Path to the output ROS2 bag file, if necessary.
+                - 'operation_params': Dictionary containing operation parameters for specific manipulations.
+                - 'external_msgs_path_ros2': Path to the directory containing external message definitions.
+                - 'operation_to_run': The name of the operation to run, which should be a method of this class.
+        """
+
+        # Assign attributes from input arguments
+        for key, value in kwargs.items():
+            setattr(self, key, value)
+
+        # Create a ROS2 bag wrapper
+        self.bag_wrapper = Ros2BagWrapper(self.input_bag, self.external_msgs_path_ros2)
+
+        # Setup dictionary to store msg instances
+        self.ros1_msg_instances = {}
+
+        # Run desired operation
+        self.run_operation()
+
+    @classmethod
+    def from_yaml(cls, yaml_path: str):
+        """
+        Create a rosbag manipulator from a YAML file.
+
+        Args:
+            yaml_path (str): Path to the YAML file containing the manipulation configuration.
+        """
+        with open(yaml_path, "r") as yaml_file:
+            yaml_dict = yaml.safe_load(yaml_file)
+            return cls(**yaml_dict)
+
+    def run_operation(self):
+        """
+        Load the operation to run from 'self.operation_to_run' and execute it.
+
+        Used Attributes:
+            self.operation_to_run (str): The name of the operation to run, which should be a method of this class.
+        """
+        function = getattr(self, self.operation_to_run)
+        function()
+
+
+    def hertz_analysis(self):
+        """
+        Analyze the hertz of various topics in a ROS2 bag file. Will output
+        a histogram of the hertz for each pair of consecutive messages
+        for the specified topic, and another histogram of the time differences.
+        """
+
+        # Extract operation specific parameters
+        topic: str = self.operation_params['hertz_analysis']['topic']
+        output_folder: str = self.operation_params['hertz_analysis']['output_folder']
+        expected_msgs: int = self.operation_params['hertz_analysis']['expected_msgs']
+        max_msgs: int = self.operation_params['hertz_analysis']['max_msgs']
+        try: robot_name: str = self.operation_params['hertz_analysis']['robot_name']
+        except: robot_name = None
+
+        self.bag_wrapper.hertz_analysis(topic, output_folder, expected_msgs, max_msgs, robot_name)
+
+    def view_imu_data(self):
+        """
+        Plot the linear acceleration data of an IMU topic in a ROS2 bag.
+        """
+
+        # Extract operation specific parameters
+        topic: str = self.operation_params['view_imu_data']['topic']
+        output_folder: str = self.operation_params['view_imu_data']['output_folder']
+        try: expected_msgs: int = self.operation_params['view_imu_data']['expected_msgs']
+        except: expected_msgs = None
+        try:
+            data_range = self.operation_params['view_imu_data']['data_range']
+            assert len(data_range) == 2
+            data_range = tuple(data_range)
+        except: data_range = None
+
+        self.bag_wrapper.view_imu_data(topic, output_folder, expected_msgs, data_range)
+
+    def downsample(self):
+        """
+        Downsample a ROS2 bag file by downsampling the frequency of specified topics. Additionally
+        includes or excludes unmentioned topics based on `include_unmentioned_topics` parameter, which
+        can be used to prune topics.
+        """
+
+        # Extract operation specific parameters
+        topic_downsample_ratios: dict = self.operation_params['downsample']['topics'].copy()
+        include_unmentioned_topics: bool = self.operation_params['downsample']['include_unmentioned_topics']
+        self.bag_wrapper.downsample(self.output_bag, topic_downsample_ratios, include_unmentioned_topics)
+
+    def crop(self):
+        """ Crop a ROS2 bag file. """
+
+        start_ts: float = self.operation_params['crop']['start_ts']
+        end_ts: float = self.operation_params['crop']['end_ts']
+        self.bag_wrapper.crop(self.output_bag, start_ts, end_ts)
+
+    def extract_odometry_to_csv(self):
+        """ Extract odometry from a ROS2 bag and save in a csv file. """
+
+        topic: str = self.operation_params['extract_odometry_to_csv']['topic']
+        output_folder: str = self.operation_params['extract_odometry_to_csv']['output_file']
+        add_noise: str = self.operation_params['extract_odometry_to_csv']['add_noise']
+        xy_noise_std_per_frame: float = self.operation_params['extract_odometry_to_csv']['xy_noise_std_per_frame']
+        z_noise_std_per_frame: str = self.operation_params['extract_odometry_to_csv']['z_noise_std_per_frame']
+        shift_position_xy: float = self.operation_params['extract_odometry_to_csv']['shift_position_xy']
+        shift_position_z: float = self.operation_params['extract_odometry_to_csv']['shift_position_z']
+
+        odom_data = OdometryData.from_ros2_bag(self.input_bag, topic)
+        if add_noise:
+            odom_data.add_folded_guassian_noise_to_position(xy_noise_std_per_frame, z_noise_std_per_frame)
+            odom_data.shift_position(shift_position_xy, shift_position_xy, shift_position_z)
+        odom_data.to_csv(output_folder)
+
+    def extract_images_to_npy(self):
+        """ Extract images from a ROS2 bag and saves them into .npy files. """
+
+        topic: str = self.operation_params['extract_images_to_npy']['topic']
+        output_folder: str = self.operation_params['extract_images_to_npy']['output_folder']
+        ImageData.from_ros2_bag(self.bag_wrapper, topic, output_folder)
+
+    def compare_timestamps_two_image_data(self):
+        """ Compare timestamps between two ImageData instances. """
+
+        data0 = ImageData.from_npy(self.operation_params['compare_timestamps_two_image_data']['folder_0'])
+        data1 = ImageData.from_npy(self.operation_params['compare_timestamps_two_image_data']['folder_1'])
+        data0.compare_timestamps(data1)
+
+    def convert_ros2_to_ros1(self):
+        """ Convert a ROS2 bag from the bag wrapper into a ROS1 bag. """
+
+        self.bag_wrapper.export_as_ros1(self.output_bag, self.external_msgs_path_ros1)
