+++ tests/test_ImageData.py
@@ -5,6 +5,7 @@ import os
 from pathlib import Path
 from rosbag_manip.data_types.ImageData import ImageData
 from rosbag_manip.rosbag.Ros2BagWrapper import Ros2BagWrapper
+from rosbag_manip.rosbag.Ros2BagWrapper import Ros2BagWrapper
 from test_utils import safe_urlretrieve
 import unittest
