+++ tests/test_OdometryData.py
@@ -28,6 +28,8 @@ class TestOdometryData(unittest.TestCase):
 
         # Load the data back again
         ros_data = OdometryData.from_ros2_bag(bag_path, '/odom')
+        ros_data.to_ROS_frame()
+        ros_data.to_ROS_frame()
 
         # Make sure this data matches what we expect
         np.testing.assert_equal(float(ros_data.timestamps[13801]), 690.100000)
